streamlit_app/requirements.txt
streamlit>=1.42.0
Authlib>=1.3.0
libsql>=0.1.11
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.22.0
finnhub-python>=2.4.20
anthropic>=0.34.0
requests>=2.32.0
lxml>=5.2.0
wordcloud>=1.9.0
matplotlib>=3.8.0
